Use less memory
Pry keeps the commands and outputs for the last 100 evals in memory. This makes notebooks with large outputs (like images) crawl to a halt. Since IRuby keeps a history of 3, pry should as well. I would love to set memory_size to 0, but I'm not sure how that would effect the IRuby::History module.